* Update the prototype of the function for keyword generation code when INN is compiled without that support.
* Add two new files for the test suite in the MANIFEST file.

* Fix two gcc 4.4 cast warnings.

+++ backends/cvtbatch.c
@@ -107,7 +107,7 @@ main(int ac, char *av[]) {
 	    case FEED_BYTESIZE:
 		if (Dirty)
 		    putchar(' ');
-		printf("%d", len);
+		printf("%lu", (unsigned long)len);
 		break;
 	    case FEED_FULLNAME:
 	    case FEED_NAME:
